MonthsSlider: move in all state and helpers
we change the MonthsSlider API to take only globalDates and setDates as
props, the slider handles all it's date->int conversion and it doesn't
spill out anymore.

this allows to reorganize App.jsx quite a bit leaving almost no date
manipulation in that file, and greatly simplifying it's state
management.

WARNING: we now use slider-container as it's name intends: to contain
the slider plus the side button. that said this commit introduces a
style change as now slider is grown by it's side button.

TODO: i'd love to get slider ticks.